test: create separate file for lifecycle hooks tests
